requirements.txt
aiogram==2.19
sqlalchemy==1.4.34
psycopg2-binary==2.9.3
python-dotenv==0.20.0
